seems to work - pulled async code out of reducer
